refactor: simplify orb slam example layout
