+++ examples/common/datasets/mod.rs
@@ -1,3 +1,3 @@
 pub mod euroc;
 
-pub use euroc::{EurocDataset, GroundTruthPose};
+pub use euroc::EurocDataset;
