+++ examples/orb_slam/main.rs
@@ -1,4 +1,4 @@
-//! ORB-SLAM example with Rerun visualization.
+//! ORB-SLAM example with minimal Rerun visualization.
 //!
 //! Runs an ORB-based SLAM pipeline on EuRoC MAV images, extracting ORB features
 //! externally and feeding them to `process_frame`.
@@ -8,30 +8,24 @@
 //! ```
 
 mod pipeline;
-mod bootstrap;
-mod tracking;
 mod utils;
 #[path = "../common/datasets/mod.rs"]
 mod datasets;
-
-use std::collections::BTreeMap;
-use std::path::PathBuf;
 
 use kornia_3d::camera::PinholeCamera;
 use kornia_3d::pose::Pose3d;
 use kornia_io::png::read_image_png_mono8;
 use kornia_slam::estimation::map_projection::MapProjectionConfig;
 use kornia_slam::estimation::two_view::TwoViewInitConfig;
-use kornia_slam::{Frame, OdometryStatus};
+use kornia_slam::Frame;
 
-use pipeline::ORBSLAMPipeline;
+use pipeline::Pipeline;
 
 use datasets::EurocDataset;
 
 use utils::{
-    RunArtifactWriter, SummaryRecord, TumPose, default_run_dir, frame_record_from_tracking_result,
-    log_ground_truth_to_rerun, log_keyframes_to_rerun, log_map_points_to_rerun,
-    log_trajectory_to_rerun, pose_to_tum, write_tum_trajectory,
+    log_camera_to_rerun, log_frame_to_rerun, log_map_points_to_rerun, log_trajectory_to_rerun,
+    trajectory_point_from_pose,
 };
 
 /// CLI arguments.
@@ -58,14 +52,6 @@ struct Args {
     #[argh(switch)]
     rerun_stream: bool,
 
-    /// output TUM trajectory file path
-    #[argh(option)]
-    output: Option<String>,
-
-    /// output directory for machine-readable run artifacts
-    #[argh(option)]
-    run_dir: Option<String>,
-
     /// skip this many initial frames
     #[argh(option, default = "0")]
     start_frame: usize,
@@ -77,7 +63,6 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
     // ── Dataset ────────────────────────────────────────────────────────────
     let dataset = EurocDataset::open(&args.data)?;
     let samples = dataset.samples();
-    let ground_truth = dataset.ground_truth();
 
     let n_frames = if args.max_frames > 0 {
         args.max_frames.min(samples.len() - args.start_frame)
@@ -90,25 +75,6 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
         args.start_frame,
         args.start_frame + n_frames
     );
-    // Temporary debugging artifacts for comparing this port against the original crate.
-    let run_dir = args
-        .run_dir
-        .as_deref()
-        .map(PathBuf::from)
-        .unwrap_or_else(|| default_run_dir("mono_vo_run"));
-    let run_name = run_dir
-        .file_name()
-        .and_then(|name| name.to_str())
-        .unwrap_or("mono_vo_run")
-        .to_string();
-    let mut run_writer = RunArtifactWriter::create(&run_dir)?;
-    let dataset_line = format!(
-        "Dataset: {} frames (processing {}..{})",
-        samples.len(),
-        args.start_frame,
-        args.start_frame + n_frames
-    );
-    run_writer.write_stdout_line(&dataset_line)?;
 
     // ── Camera (EuRoC cam0 intrinsics) ─────────────────────────────────────
     let camera = PinholeCamera {
@@ -122,16 +88,18 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
         p2: 1.76187114e-05,
     };
 
-    // ── ORB detector (used externally before feeding ORBSLAMPipeline) ─────
+    // ── ORB detector (used externally before feeding Pipeline) ─────────────
     let detector = kornia_imgproc::features::OrbDetector {
         n_keypoints: args.n_keypoints,
         ..Default::default()
     };
 
     // ── SLAM config & system ───────────────────────────────────────────────
-    let two_view_init_config = TwoViewInitConfig::default();
+    let mut two_view_init_config = TwoViewInitConfig::default();
+    two_view_init_config.estimation_config.triangulation.max_midpoint_gap = 0.25;
+    two_view_init_config.estimation_config.triangulation.max_reprojection_error = 3.0;
     let map_projection_config = MapProjectionConfig::default();
-    let mut system = ORBSLAMPipeline::new(camera, two_view_init_config, map_projection_config);
+    let mut system = Pipeline::new(camera, two_view_init_config, map_projection_config);
 
     // ── Rerun ──────────────────────────────────────────────────────────────
     let rec = if args.rerun_stream {
@@ -152,10 +120,7 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
     };
 
     // ── Main loop ──────────────────────────────────────────────────────────
-    let mut trajectory: Vec<TumPose> = Vec::with_capacity(n_frames);
-    let mut keyframe_poses: Vec<(usize, kornia_3d::pose::Pose3d)> = Vec::new();
-    let mut keyframe_tum: Vec<TumPose> = Vec::new();
-    let mut status_counts: BTreeMap<String, usize> = BTreeMap::new();
+    let mut trajectory: Vec<[f32; 3]> = Vec::with_capacity(n_frames);
 
     for (i, sample) in samples
         .iter()
@@ -172,14 +137,13 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
             height: gray_u8.height() as f64,
         };
         let gray_f32 = {
-            let inner = gray_u8.into_inner();
             let mut dst = kornia_image::Image::from_size_val(
-                inner.size(),
+                gray_u8.size(),
                 0.0f32,
                 kornia_tensor::CpuAllocator,
             )
             .unwrap();
-            inner
+            gray_u8
                 .as_slice()
                 .iter()
                 .zip(dst.as_slice_mut())
@@ -189,6 +153,9 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
 
         // Extract ORB features.
         let features = detector.detect_and_extract(&gray_f32)?;
+        if let Some(ref rec) = rec {
+            log_frame_to_rerun(rec, &gray_u8, &features.keypoints_xy);
+        }
 
         // Run SLAM.
         let frame = Frame::new(idx, features, Pose3d::IDENTITY, image_size);
@@ -202,61 +169,19 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
             result.status, keyframe_idx, map_point_count,
         );
         eprintln!("{status_line}");
-        run_writer.write_stdout_line(&status_line)?;
-        *status_counts
-            .entry(format!("{:?}", result.status))
-            .or_insert(0) += 1;
 
         // Collect trajectory.
-        let tum_pose = pose_to_tum(sample.timestamp_sec, &result.pose_world_to_cam);
-        trajectory.push(tum_pose);
-
-        // Track keyframes for visualization.
-        if result.status == OdometryStatus::KeyframeAccepted {
-            keyframe_poses.push((keyframe_idx, result.pose_world_to_cam));
-            keyframe_tum.push(tum_pose);
-        }
-        run_writer.write_frame(&frame_record_from_tracking_result(
-            idx,
-            sample.timestamp_sec,
-            keyframe_idx,
-            map_point_count,
-            &result,
-        ))?;
+        trajectory.push(trajectory_point_from_pose(&result.pose_world_to_cam));
 
         // Rerun logging.
         if let Some(ref rec) = rec {
-            log_trajectory_to_rerun(rec, &trajectory, &result.pose_world_to_cam);
+            log_trajectory_to_rerun(rec, &trajectory);
+            log_camera_to_rerun(rec, &result.pose_world_to_cam);
             log_map_points_to_rerun(rec, system.map_points());
-            log_keyframes_to_rerun(rec, &keyframe_poses);
-            log_ground_truth_to_rerun(rec, ground_truth, sample.timestamp_sec);
         }
     }
 
-    // ── Write output ───────────────────────────────────────────────────────
-    if let Some(ref out) = args.output {
-        write_tum_trajectory(&PathBuf::from(out), &trajectory)?;
-        eprintln!("Wrote {}-pose trajectory to {out}", trajectory.len());
-        run_writer.write_stdout_line(&format!(
-            "Wrote {}-pose trajectory to {out}",
-            trajectory.len()
-        ))?;
-    }
-
-    let summary = SummaryRecord {
-        dataset_path: args.data.clone(),
-        run_name,
-        frame_count: trajectory.len(),
-        status_counts,
-        final_map_points: system.map_points().len(),
-    };
-    run_writer.write_summary(&summary, &trajectory, &keyframe_tum)?;
-
     let final_line = format!("Done. Final map: {} points", system.map_points().len());
     eprintln!("{final_line}");
-    run_writer.write_stdout_line(&final_line)?;
-    let run_dir_line = format!("Wrote run artifacts to {}", run_dir.display());
-    eprintln!("{run_dir_line}");
-    run_writer.write_stdout_line(&run_dir_line)?;
     Ok(())
 }
