+++ src/lib.rs
@@ -2,7 +2,7 @@
 
 pub mod frame;
 pub mod estimation;
-pub mod mapping;
+pub mod map;
 pub mod odometry;
 
 pub use frame::{Frame, OrbFeatures};
