+++ src/map.rs
@@ -1,8 +1,36 @@
 //! Map: keyframes, map points, local map selection, and culling.
+//!
+//! ```text
+//!    +--------+
+//!    | Frame  |
+//!    +--------+
+//!         |
+//!         v
+//!    +----------------------+
+//!    | Keyframe             |
+//!    | frame + desc -> mp   |
+//!    +----------------------+
+//!         |
+//!         v
+//!    +----------------------+
+//!    | Map                  |
+//!    | keyframes + points   |
+//!    +----------------------+
+//!
+//!    ops:
+//!      * upsert_keyframe
+//!      * push_map_point
+//!      * build_local_map_points
+//!      * cull_map_points
+//!      * run_local_ba
+//! ```
 
 use std::collections::{HashMap, HashSet};
 
 use kornia_algebra::Vec3F64;
+use kornia_3d::ba::{self, BaObservation, BaParams};
+use kornia_3d::camera::PinholeCamera;
+use kornia_3d::pose::Pose3d;
 
 use crate::frame::Frame;
 
@@ -278,6 +306,102 @@ pub fn cull_map_points(map: &mut Map) {
                     }
                 }
             }
+        }
+    }
+}
+
+/// Run local bundle adjustment over recent keyframes and their observed map points.
+///
+/// Collects the last N active keyframes, gathers observations (undistorting keypoints
+/// via camera), calls `kornia_3d::ba::bundle_adjust`, and writes back optimized poses
+/// and point positions into the map.
+pub fn run_local_ba(map: &mut Map, camera: &PinholeCamera) {
+    const MAX_ACTIVE_KFS: usize = 3;
+    const MIN_OBSERVATIONS: usize = 8;
+
+    let n_kfs = map.keyframes().len();
+    if n_kfs < 2 {
+        return;
+    }
+
+    let active_start = n_kfs.saturating_sub(MAX_ACTIVE_KFS);
+
+    let mut mp_set: HashSet<usize> = HashSet::new();
+    for kf in &map.keyframes()[active_start..] {
+        for mp_idx in kf.map_point_by_desc_idx.iter().flatten() {
+            if let Some(mp) = map.map_points().get(*mp_idx) {
+                if !mp.culled {
+                    mp_set.insert(*mp_idx);
+                }
+            }
+        }
+    }
+    if mp_set.is_empty() {
+        return;
+    }
+
+    let mut mp_global_indices: Vec<usize> = mp_set.iter().copied().collect();
+    mp_global_indices.sort_unstable();
+
+    let mp_global_to_local: HashMap<usize, usize> = mp_global_indices
+        .iter()
+        .enumerate()
+        .map(|(local, &global)| (global, local))
+        .collect();
+
+    let points: Vec<Vec3F64> = mp_global_indices
+        .iter()
+        .map(|&idx| map.map_points()[idx].position)
+        .collect();
+
+    let poses: Vec<Pose3d> = map
+        .keyframes()
+        .iter()
+        .map(|kf| kf.frame.pose_world_to_cam)
+        .collect();
+
+    let mut observations = Vec::new();
+    for (kf_idx, kf) in map.keyframes().iter().enumerate() {
+        let is_fixed = kf_idx < active_start;
+        for (desc_idx, mp_opt) in kf.map_point_by_desc_idx.iter().enumerate() {
+            if let Some(mp_idx) = mp_opt {
+                let Some(&point_idx) = mp_global_to_local.get(mp_idx) else {
+                    continue;
+                };
+                if let Some(kp) = kf.frame.features.keypoints_xy.get(desc_idx) {
+                    let p = camera.undistort(kp[0] as f64, kp[1] as f64);
+                    observations.push(BaObservation {
+                        pose_idx: kf_idx,
+                        point_idx,
+                        pixel: [p.x as f32, p.y as f32],
+                        fixed_pose: is_fixed,
+                    });
+                }
+            }
+        }
+    }
+
+    if observations.len() < MIN_OBSERVATIONS {
+        return;
+    }
+
+    let ba_result =
+        match ba::bundle_adjust(&poses, &points, &observations, camera, &BaParams::default()) {
+            Ok(r) => r,
+            Err(_) => return,
+        };
+
+    let keyframes = map.keyframes_mut();
+    for (kf_idx, pose) in ba_result.poses.iter().enumerate() {
+        if kf_idx >= active_start {
+            keyframes[kf_idx].frame.pose_world_to_cam = *pose;
+        }
+    }
+
+    let map_points = map.map_points_mut();
+    for (local_idx, &global_idx) in mp_global_indices.iter().enumerate() {
+        if let Some(mp) = map_points.get_mut(global_idx) {
+            mp.position = ba_result.points[local_idx];
         }
     }
 }
